pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name            = "osint-tool"
version         = "3.0.0"
description     = "Comprehensive Open Source Intelligence framework — 27 modules"
readme          = "README.md"
license         = { file = "LICENSE" }
requires-python = ">=3.9"

keywords = [
    "osint", "intelligence", "malware", "darkweb", "certification", "security", "recon", "reconnaissance",
    "forensics", "threat-intelligence", "penetration-testing",
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "rich>=13.7.0",
    "dnspython>=2.4.2",
    "python-whois>=0.9.4",
    "Pillow>=10.2.0",
    "exifread>=3.0.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "phonenumbers>=8.13.27",
    "matplotlib>=3.8.2",
    "folium>=0.15.1",
    "pysocks>=1.7.1",
    "stem>=1.8.2",
    "python-dotenv>=1.0.1",
    "shodan>=1.31.0",
    "pyvis>=0.3.2",
    "networkx>=3.2.1",
    "schedule>=1.2.1",
    "tqdm>=4.66.1",
    "colorama>=0.4.6",
    "fake-useragent>=1.4.0",
    "geopy>=2.4.1",
    "pandas>=2.2.0",
    "jinja2>=3.1.3",
    "fpdf2>=2.7.9",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "flake8>=7.0.0",
    "bandit>=1.7.7",
    "mypy>=1.8.0",
    "isort>=5.13.2",
    "pre-commit>=3.6.0",
]

[project.scripts]
osint-tool = "main:cli_mode"

[project.urls]
Homepage       = "https://github.com/YOUR_USERNAME/osint-tool"
Documentation  = "https://github.com/YOUR_USERNAME/osint-tool#readme"
Repository     = "https://github.com/YOUR_USERNAME/osint-tool"
"Bug Tracker"  = "https://github.com/YOUR_USERNAME/osint-tool/issues"
Changelog      = "https://github.com/YOUR_USERNAME/osint-tool/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["modules*", "reporting*", "utils*"]

# ── Black ─────────────────────────────────────────────────
[tool.black]
line-length    = 100
target-version = ["py39", "py310", "py311", "py312"]
skip-string-normalization = false

# ── isort ─────────────────────────────────────────────────
[tool.isort]
profile      = "black"
line_length  = 100

# ── flake8 (via setup.cfg fallback) ───────────────────────
# flake8 doesn't support pyproject.toml natively; see .flake8

# ── pytest ────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts      = "-v --tb=short"

# ── mypy ──────────────────────────────────────────────────
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ── bandit ────────────────────────────────────────────────
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]  # assert used in tests
